Rename package and classes;
Change name package from manager to service;
Change name package from serviceByRepo to service.repo;
Change name interface from ServiceByRepo to ServiceRepo;
Change name classes in service.repo package from **ServiceByRepoImp to **ServiceRepoImp
