+++ src/main/java/com/CezaryZal/api/service/calculation/result/Result.java
@@ -1,7 +1,7 @@
-package com.CezaryZal.api.manager.calculation.result;
+package com.CezaryZal.api.service.calculation.result;
 
-import com.CezaryZal.api.manager.calculation.deviation.LowerDeviation;
-import com.CezaryZal.api.manager.calculation.deviation.UpperDeviation;
+import com.CezaryZal.api.service.calculation.deviation.LowerDeviation;
+import com.CezaryZal.api.service.calculation.deviation.UpperDeviation;
 import com.CezaryZal.api.model.ParsedInputDimension;
 import com.CezaryZal.api.model.ValuesFromRepoByInputDimension;
 import com.CezaryZal.api.model.dto.DimensionDTO;
